+++ packages/chain-registry/src/devnet/assets.ts
@@ -4,6 +4,7 @@ import * as _bitcannadevnet2 from './bitcannadevnet2';
 import * as _celestiatestnet2 from './celestiatestnet2';
 import * as _impacthubdevnet from './impacthubdevnet';
 import * as _kyvedevnet from './kyvedevnet';
+import * as _neuradevnet from './neuradevnet';
 import * as _nibirudevnet from './nibirudevnet';
 import * as _nibirudevnet2 from './nibirudevnet2';
 import * as _seidevnet3 from './seidevnet3';
@@ -13,6 +14,7 @@ const assets: AssetList[] = [
   _celestiatestnet2.assets,
   _impacthubdevnet.assets,
   _kyvedevnet.assets,
+  _neuradevnet.assets,
   _nibirudevnet.assets,
   _nibirudevnet2.assets,
   _seidevnet3.assets