+++ packages/chain-registry/src/mainnet/aura/chain.ts
@@ -6,10 +6,11 @@ const info: Chain = {
   network_type: 'mainnet',
   website: 'https://aura.network/',
   pretty_name: 'Aura Network',
-  chain_id: 'xstaxy-1',
+  chain_id: 'aura_6322-2',
   bech32_prefix: 'aura',
   daemon_name: 'aurad',
   node_home: '$HOME/.aura',
+  pre_fork_chain_name: 'aura1',
   key_algos: ['secp256k1'],
   slip44: 118,
   fees: {
@@ -30,9 +31,9 @@ const info: Chain = {
     }
   },
   codebase: {
-    cosmos_sdk_version: '0.47.5',
+    cosmos_sdk_version: 'evmos/cosmos-sdk v0.47.8-evmos',
     cosmwasm_enabled: true,
-    cosmwasm_version: '0.42.0'
+    cosmwasm_version: 'v0.42.0'
   },
   logo_URIs: {
     png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/aura/images/Aura-logo-2.2.png',
@@ -254,8 +255,8 @@ const info: Chain = {
     {
       kind: 'aurascan',
       url: 'https://aurascan.io',
-      tx_page: 'https://aurascan.io/transaction/${txHash}',
-      account_page: 'https://aurascan.io/account/${accountAddress}'
+      tx_page: 'https://aurascan.io/tx/${txHash}',
+      account_page: 'https://aurascan.io/address/${accountAddress}'
     },
     {
       kind: '𝐥𝐞𝐬𝐧𝐢𝐤 | 𝐔𝐓𝐒𝐀 Explorer',