+++ legacy/juno/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.63.0 (2024-06-04)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
+## 1.62.12 (2024-06-04)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
+## [1.62.11](https://github.com/cosmology-tech/chain-registry/compare/@chain-registry/juno@1.62.10...@chain-registry/juno@1.62.11) (2024-06-03)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
+## [1.62.10](https://github.com/cosmology-tech/chain-registry/compare/@chain-registry/juno@1.62.9...@chain-registry/juno@1.62.10) (2024-06-02)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
+## 1.62.9 (2024-06-01)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
+## 1.62.8 (2024-05-31)
+
+**Note:** Version bump only for package @chain-registry/juno
+
+
+
+
+
 ## 1.62.7 (2024-05-30)
 
 **Note:** Version bump only for package @chain-registry/juno