+++ workflows/cli/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.48.1](https://github.com/cosmology-tech/chain-registry/compare/@chain-registry/cli@1.48.0...@chain-registry/cli@1.48.1) (2024-06-04)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
+# 1.48.0 (2024-06-04)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
+## 1.47.10 (2024-06-04)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
+## 1.47.9 (2024-06-01)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
+## 1.47.8 (2024-05-31)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
 ## 1.47.7 (2024-05-30)
 
 **Note:** Version bump only for package @chain-registry/cli